Move query types.
